+++ glade_forms/lib/src/core/input/glade_input.dart
@@ -2,35 +2,22 @@ import 'dart:math';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/widgets.dart';
-import 'package:glade_forms/src/converters/glade_type_converters.dart';
 import 'package:glade_forms/src/core/changes_info.dart';
 import 'package:glade_forms/src/core/error/convert_error.dart';
 import 'package:glade_forms/src/core/error/error_translator.dart';
 import 'package:glade_forms/src/core/input_dependencies.dart';
 import 'package:glade_forms/src/core/string_to_type_converter.dart';
-import 'package:glade_forms/src/utils/type_helper.dart';
 import 'package:glade_forms/src/model/glade_model.dart';
-import 'package:glade_forms/src/validator/specialized/date_time_validator.dart';
+import 'package:glade_forms/src/utils/type_helper.dart';
 import 'package:glade_forms/src/validator/validator.dart';
 import 'package:glade_forms/src/validator/validator_result.dart';
 import 'package:meta/meta.dart';
 
 typedef ValueComparator<T> = bool Function(T? initial, T? value);
-typedef ValidatorFactory<T> = ValidatorInstance<T> Function(GladeValidator<T> v);
-typedef StringValidatorFactory = ValidatorInstance<String> Function(StringValidator validator);
-typedef IntValidatorFactory = ValidatorInstance<int> Function(IntValidator validator);
-typedef DateTimeValidatorFactory = ValidatorInstance<DateTime> Function(DateTimeValidator validator);
-typedef DateTimeValidatorFactoryNullable = ValidatorInstance<DateTime?> Function(DateTimeValidatorNullable validator);
+
 typedef OnChange<T> = void Function(ChangesInfo<T> info);
 typedef OnDependencyChange = void Function(List<String> updateInputKeys);
 typedef ValueTransform<T> = T Function(T input);
-
-typedef StringInput = GladeInput<String>;
-typedef StringInputNullable = GladeInput<String?>;
-typedef IntInput = GladeInput<int>;
-typedef BooleanInput = GladeInput<bool>;
-typedef DateTimeInput = GladeInput<DateTime>;
-typedef DateTimeNullableInput = GladeInput<DateTime?>;
 
 class GladeInput<T> {
   /// Compares initial and current value.
@@ -153,31 +140,32 @@ class GladeInput<T> {
     _bindedModel?.notifyInputUpdated(this);
   }
 
-  GladeInput._({
-    required T value,
+  @internal
+  GladeInput.internalCreate({
     required this.validatorInstance,
-    required bool isPure,
-    required this.valueComparator,
-    required String? inputKey,
-    required this.translateError,
-    required this.stringToValueConverter,
-    required InputDependenciesFactory? dependenciesFactory,
-    required this.defaultTranslations,
-    required this.onChange,
-    required this.onDependencyChange,
-    required ValueTransform<T>? valueTransform,
+    String? inputKey,
+    T? value,
     T? initialValue,
+    bool isPure = true,
+    this.translateError,
+    this.valueComparator,
+    this.stringToValueConverter,
+    InputDependenciesFactory? dependencies,
+    this.onChange,
+    this.onDependencyChange,
     TextEditingController? textEditingController,
     bool useTextEditingController = false,
+    ValueTransform<T>? valueTransform,
+    this.defaultTranslations,
     this.trackUnchanged = true,
   })  : assert(
-          dependenciesFactory == null || (onDependencyChange != null),
-          'When dependencies are provided, provide onDependencyChange as well',
+          value != null || initialValue != null || TypeHelper.typeIsNullable<T>(),
+          'If type is not nullable, at least one of value or initialValue must be set (affected input: $inputKey)',
         ),
         _isPure = isPure,
-        _value = value,
+        _value = (value ?? initialValue) as T,
         _initialValue = initialValue,
-        dependenciesFactory = dependenciesFactory ?? (() => []),
+        dependenciesFactory = dependencies ?? (() => []),
         inputKey = inputKey ?? '__${T.runtimeType}__${Random().nextInt(100000000)}',
         _valueTransform = valueTransform,
         _textEditingController = textEditingController ??
@@ -190,6 +178,7 @@ class GladeInput<T> {
                     },
                   )
                 : null),
+
         // ignore: avoid_bool_literals_in_conditional_expressions, cant be simplified.
         _useTextEditingController = textEditingController != null ? true : useTextEditingController {
     validatorInstance.bindInput(this);
@@ -202,21 +191,13 @@ class GladeInput<T> {
   /// At least one of [value] or [initialValue] MUST be set.
   factory GladeInput.create({
     String? inputKey,
-
-    /// Sets current value of input.
-    /// When value is null, it is set to [initialValue].
     T? value,
-
-    /// Initial value when GenericInput is created.
-    ///
-    /// This value can potentially differ from [value].
-    /// Used for computing [isUnchanged].
     T? initialValue,
     ValidatorFactory<T>? validator,
-    bool pure = true,
+    bool isPure = true,
     ErrorTranslator<T>? translateError,
     ValueComparator<T>? valueComparator,
-    StringToTypeConverter<T>? valueConverter,
+    StringToTypeConverter<T>? stringToValueConverter,
     InputDependenciesFactory? dependencies,
     OnChange<T>? onChange,
     OnDependencyChange? onDependencyChange,
@@ -225,33 +206,25 @@ class GladeInput<T> {
     ValueTransform<T>? valueTransform,
     DefaultTranslations? defaultTranslations,
     bool trackUnchanged = true,
-  }) {
-    assert(
-      value != null || initialValue != null || TypeHelper.typeIsNullable<T>(),
-      'If type is not nullable, at least one of value or initialValue must be set',
-    );
-
-    final validatorInstance = validator?.call(GladeValidator()) ?? GladeValidator<T>().build();
-
-    return GladeInput._(
-      value: (value ?? initialValue) as T,
-      isPure: pure,
-      validatorInstance: validatorInstance,
-      initialValue: initialValue,
-      translateError: translateError,
-      valueComparator: valueComparator,
-      inputKey: inputKey,
-      stringToValueConverter: valueConverter,
-      dependenciesFactory: dependencies,
-      onChange: onChange,
-      onDependencyChange: onDependencyChange,
-      textEditingController: textEditingController,
-      useTextEditingController: useTextEditingController,
-      valueTransform: valueTransform,
-      defaultTranslations: defaultTranslations,
-      trackUnchanged: trackUnchanged,
-    );
-  }
+  }) =>
+      GladeInput.internalCreate(
+        validatorInstance: validator?.call(GladeValidator()) ?? GladeValidator<T>().build(),
+        inputKey: inputKey,
+        value: value,
+        initialValue: initialValue,
+        isPure: isPure,
+        translateError: translateError,
+        valueComparator: valueComparator,
+        stringToValueConverter: stringToValueConverter,
+        dependencies: dependencies,
+        onChange: onChange,
+        onDependencyChange: onDependencyChange,
+        textEditingController: textEditingController,
+        useTextEditingController: useTextEditingController,
+        valueTransform: valueTransform,
+        defaultTranslations: defaultTranslations,
+        trackUnchanged: trackUnchanged,
+      );
 
   ///
   /// Useful for input which allows null value without additional validations.
@@ -265,7 +238,7 @@ class GladeInput<T> {
     ErrorTranslator<T>? translateError,
     DefaultTranslations? defaultTranslations,
     ValueComparator<T>? valueComparator,
-    StringToTypeConverter<T>? valueConverter,
+    StringToTypeConverter<T>? stringToValueConverter,
     InputDependenciesFactory? dependencies,
     OnChange<T>? onChange,
     OnDependencyChange? onDependencyChange,
@@ -281,9 +254,9 @@ class GladeInput<T> {
         translateError: translateError,
         defaultTranslations: defaultTranslations,
         valueComparator: valueComparator,
-        valueConverter: valueConverter,
+        stringToValueConverter: stringToValueConverter,
         inputKey: inputKey,
-        pure: pure,
+        isPure: pure,
         dependencies: dependencies,
         onChange: onChange,
         onDependencyChange: onDependencyChange,
@@ -304,7 +277,7 @@ class GladeInput<T> {
     ErrorTranslator<T>? translateError,
     DefaultTranslations? defaultTranslations,
     ValueComparator<T>? valueComparator,
-    StringToTypeConverter<T>? valueConverter,
+    StringToTypeConverter<T>? stringToValueConverter,
     InputDependenciesFactory? dependencies,
     OnChange<T>? onChange,
     OnDependencyChange? onDependencyChange,
@@ -320,9 +293,9 @@ class GladeInput<T> {
         translateError: translateError,
         defaultTranslations: defaultTranslations,
         valueComparator: valueComparator,
-        valueConverter: valueConverter,
+        stringToValueConverter: stringToValueConverter,
         inputKey: inputKey,
-        pure: pure,
+        isPure: pure,
         dependencies: dependencies,
         onChange: onChange,
         onDependencyChange: onDependencyChange,
@@ -335,200 +308,6 @@ class GladeInput<T> {
   @internal
   // ignore: use_setters_to_change_properties, as method.
   void bindToModel(GladeModel model) => _bindedModel = model;
-
-  static IntInput intInput({
-    required int value,
-    String? inputKey,
-    int? initialValue,
-    IntValidatorFactory? validator,
-    bool pure = true,
-    ErrorTranslator<int>? translateError,
-    DefaultTranslations? defaultTranslations,
-    ValueComparator<int>? valueComparator,
-    InputDependenciesFactory? dependencies,
-    OnChange<int>? onChange,
-    OnDependencyChange? onDependencyChange,
-    TextEditingController? textEditingController,
-    bool useTextEditingController = false,
-    ValueTransform<int>? valueTransform,
-    bool trackUnchanged = true,
-  }) {
-    final validatorInstance = validator?.call(IntValidator()) ?? IntValidator().build();
-
-    return GladeInput._(
-      value: value,
-      initialValue: initialValue ?? value,
-      validatorInstance: validatorInstance,
-      isPure: pure,
-      translateError: translateError,
-      defaultTranslations: defaultTranslations,
-      valueComparator: valueComparator,
-      inputKey: inputKey,
-      dependenciesFactory: dependencies,
-      stringToValueConverter: GladeTypeConverters.intConverter,
-      onChange: onChange,
-      onDependencyChange: onDependencyChange,
-      textEditingController: textEditingController,
-      useTextEditingController: useTextEditingController,
-      valueTransform: valueTransform,
-      trackUnchanged: trackUnchanged,
-    );
-  }
-
-  static BooleanInput boolInput({
-    required bool value,
-    String? inputKey,
-    bool? initialValue,
-    ValidatorFactory<bool>? validator,
-    bool pure = true,
-    ErrorTranslator<bool>? translateError,
-    DefaultTranslations? defaultTranslations,
-    ValueComparator<bool>? valueComparator,
-    InputDependenciesFactory? dependencies,
-    OnChange<bool>? onChange,
-    OnDependencyChange? onDependencyChange,
-    TextEditingController? textEditingController,
-    bool useTextEditingController = false,
-    ValueTransform<bool>? valueTransform,
-    bool trackUnchanged = true,
-  }) =>
-      GladeInput.create(
-        value: value,
-        initialValue: initialValue ?? value,
-        validator: validator,
-        pure: pure,
-        translateError: translateError,
-        defaultTranslations: defaultTranslations,
-        valueComparator: valueComparator,
-        inputKey: inputKey,
-        dependencies: dependencies,
-        valueConverter: GladeTypeConverters.boolConverter,
-        onChange: onChange,
-        onDependencyChange: onDependencyChange,
-        textEditingController: textEditingController,
-        useTextEditingController: useTextEditingController,
-        valueTransform: valueTransform,
-        trackUnchanged: trackUnchanged,
-      );
-
-  static DateTimeInput dateTimeInput({
-    required DateTime value,
-    String? inputKey,
-    DateTime? initialValue,
-    DateTimeValidatorFactory? validator,
-    bool pure = true,
-    ErrorTranslator<DateTime>? translateError,
-    DefaultTranslations? defaultTranslations,
-    ValueComparator<DateTime>? valueComparator,
-    InputDependenciesFactory? dependencies,
-    OnChange<DateTime>? onChange,
-    OnDependencyChange? onDependencyChange,
-    TextEditingController? textEditingController,
-    bool useTextEditingController = false,
-    ValueTransform<DateTime>? valueTransform,
-    bool trackUnchanged = true,
-  }) {
-    final validatorInstance = validator?.call(DateTimeValidator()) ?? DateTimeValidator().build();
-
-    return GladeInput._(
-      value: value,
-      initialValue: initialValue ?? value,
-      validatorInstance: validatorInstance,
-      isPure: pure,
-      translateError: translateError,
-      defaultTranslations: defaultTranslations,
-      valueComparator: valueComparator,
-      inputKey: inputKey,
-      dependenciesFactory: dependencies,
-      stringToValueConverter: GladeTypeConverters.dateTimeIso8601,
-      onChange: onChange,
-      onDependencyChange: onDependencyChange,
-      textEditingController: textEditingController,
-      useTextEditingController: useTextEditingController,
-      valueTransform: valueTransform,
-      trackUnchanged: trackUnchanged,
-    );
-  }
-
-  static DateTimeNullableInput dateTimeInputNullable({
-    required DateTime? value,
-    String? inputKey,
-    DateTime? initialValue,
-    DateTimeValidatorFactoryNullable? validator,
-    bool pure = true,
-    ErrorTranslator<DateTime?>? translateError,
-    DefaultTranslations? defaultTranslations,
-    ValueComparator<DateTime?>? valueComparator,
-    InputDependenciesFactory? dependencies,
-    OnChange<DateTime?>? onChange,
-    OnDependencyChange? onDependencyChange,
-    TextEditingController? textEditingController,
-    bool useTextEditingController = false,
-    ValueTransform<DateTime?>? valueTransform,
-    bool trackUnchanged = true,
-  }) {
-    final validatorInstance = validator?.call(DateTimeValidatorNullable()) ?? DateTimeValidatorNullable().build();
-
-    return GladeInput._(
-      value: value,
-      initialValue: initialValue ?? value,
-      validatorInstance: validatorInstance,
-      isPure: pure,
-      translateError: translateError,
-      defaultTranslations: defaultTranslations,
-      valueComparator: valueComparator,
-      inputKey: inputKey,
-      dependenciesFactory: dependencies,
-      stringToValueConverter: GladeTypeConverters.dateTimeIso8601,
-      onChange: onChange,
-      onDependencyChange: onDependencyChange,
-      textEditingController: textEditingController,
-      useTextEditingController: useTextEditingController,
-      valueTransform: valueTransform,
-      trackUnchanged: trackUnchanged,
-    );
-  }
-
-  static StringInput stringInput({
-    String? inputKey,
-    String? value,
-    String? initialValue,
-    StringValidatorFactory? validator,
-    bool pure = true,
-    ErrorTranslator<String>? translateError,
-    DefaultTranslations? defaultTranslations,
-    InputDependenciesFactory? dependencies,
-    OnChange<String>? onChange,
-    OnDependencyChange? onDependencyChange,
-    TextEditingController? textEditingController,
-    bool useTextEditingController = true,
-    bool isRequired = true,
-    ValueTransform<String>? valueTransform,
-    ValueComparator<String>? valueComparator,
-    bool trackUnchanged = true,
-  }) {
-    final requiredInstance = validator?.call(StringValidator()..notEmpty()) ?? (StringValidator()..notEmpty()).build();
-    final optionalInstance = validator?.call(StringValidator()) ?? StringValidator().build();
-
-    return GladeInput._(
-      value: value ?? initialValue ?? '',
-      isPure: pure,
-      initialValue: initialValue ?? '',
-      validatorInstance: isRequired ? requiredInstance : optionalInstance,
-      translateError: translateError,
-      defaultTranslations: defaultTranslations,
-      inputKey: inputKey,
-      dependenciesFactory: dependencies,
-      onChange: onChange,
-      onDependencyChange: onDependencyChange,
-      textEditingController: textEditingController,
-      useTextEditingController: useTextEditingController,
-      valueComparator: valueComparator,
-      stringToValueConverter: null,
-      valueTransform: valueTransform,
-      trackUnchanged: trackUnchanged,
-    );
-  }
 
   // *
   // * Public methods
@@ -675,7 +454,7 @@ class GladeInput<T> {
     ValueComparator<T>? valueComparator,
     ValidatorInstance<T>? validatorInstance,
     StringToTypeConverter<T>? stringToValueConverter,
-    InputDependenciesFactory? dependenciesFactory,
+    InputDependenciesFactory? dependencies,
     T? initialValue,
     ErrorTranslator<T>? translateError,
     T? value,
@@ -689,12 +468,12 @@ class GladeInput<T> {
     ValueTransform<T>? valueTransform,
     bool? trackUnchanged,
   }) {
-    return GladeInput._(
+    return GladeInput.internalCreate(
       value: value ?? this.value,
       valueComparator: valueComparator ?? this.valueComparator,
       validatorInstance: validatorInstance ?? this.validatorInstance,
       stringToValueConverter: stringToValueConverter ?? this.stringToValueConverter,
-      dependenciesFactory: dependenciesFactory ?? this.dependenciesFactory,
+      dependencies: dependencies ?? this.dependenciesFactory,
       inputKey: inputKey ?? this.inputKey,
       initialValue: initialValue ?? this.initialValue,
       translateError: translateError ?? this.translateError,