+++ storybook/lib/usecases/dependencies/checkbox_dependency_change.dart
@@ -11,38 +11,31 @@ abstract final class _ErrorKeys {
 }
 
 class AgeRestrictedModel extends GladeModel {
-  late StringInput nameInput;
-  late GladeInput<int> ageInput;
-  late GladeInput<bool> vipInput;
+  late GladeStringInput nameInput;
+  late GladeIntInput ageInput;
+  late GladeBoolInput vipInput;
 
   @override
   List<GladeInput<Object?>> get inputs => [nameInput, ageInput, vipInput];
 
   @override
   void initialize() {
-    nameInput = GladeInput.stringInput(
+    nameInput = GladeStringInput(
       inputKey: 'name-input',
       defaultTranslations: DefaultTranslations(
         defaultValueIsNullOrEmptyMessage: LocaleKeys.empty.tr(),
       ),
     );
-    ageInput = GladeInput.create(
+    ageInput = GladeIntInput(
       validator: (v) => (v
             ..notNull()
-            ..satisfy(
-              (value) {
-                if (!vipInput.value) {
-                  return true;
-                }
-
-                return value >= 18;
-              },
-              devError: (_) => 'When VIP enabled you must be at least 18 years old.',
+            ..isMin(
+              min: 18,
+              shouldValidate: (_) => vipInput.value,
               key: _ErrorKeys.ageRestriction,
             ))
           .build(),
       value: 0,
-      valueConverter: GladeTypeConverters.intConverter,
       inputKey: 'age-input',
       translateError: (error, key, devMessage, dependencies) {
         if (key == _ErrorKeys.ageRestriction) return LocaleKeys.ageRestriction_under18.tr();
@@ -53,7 +46,7 @@ class AgeRestrictedModel extends GladeModel {
       },
       useTextEditingController: true,
     );
-    vipInput = GladeInput.create(
+    vipInput = GladeBoolInput(
       validator: (v) => (v..notNull()).build(),
       value: false,
       inputKey: 'vip-input',